refactor: model 폴더 domain, dto, generator 구분
