--- src/main/java/lotto/model/LottoRankCount.java
@@ -1,5 +1,0 @@
-package lotto.model;
-
-public record LottoRankCount(LottoRank rank, int count) {
-}
-
