+++ src/main/java/lotto/model/domain/LottoMachine.java
@@ -1,7 +1,8 @@
-package lotto.model;
+package lotto.model.domain;
 
 import java.util.ArrayList;
 import java.util.List;
+import lotto.model.generator.NumberGenerator;
 
 public class LottoMachine {
 
@@ -11,7 +12,7 @@ public class LottoMachine {
         this.numberGenerator = numberGenerator;
     }
 
-    public List<Lotto> exchangeLotto(LottoTickets lottoTickets) {
+    public List<Lotto> generateLottos(LottoTickets lottoTickets) {
         ArrayList<Lotto> lottos = new ArrayList<>();
         while (lottoTickets.hasTicketCount()) {
             lottos.add(new Lotto(numberGenerator.generateNumbersInRange()));
