+++ src/main/java/lotto/model/domain/LottoRank.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.domain;
 
 import java.util.Arrays;
