+++ src/main/java/lotto/model/domain/WinningNumbers.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.domain;
 
 import java.util.Arrays;
 import java.util.HashSet;
