+++ src/main/java/lotto/model/dto/LottoRankCount.java
@@ -1,0 +1,7 @@
+package lotto.model.dto;
+
+import lotto.model.domain.LottoRank;
+
+public record LottoRankCount(LottoRank rank, int count) {
+}
+
