+++ src/main/java/lotto/model/dto/LottoStatisticsDto.java
@@ -1,8 +1,9 @@
-package lotto.model;
+package lotto.model.dto;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import lotto.model.domain.LottoRank;
 
 public class LottoStatisticsDto {
     private final List<LottoRankCount> sortedRankCounts;
