+++ src/main/java/lotto/model/generator/NumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.generator;
 
 import java.util.List;
