+++ src/main/java/lotto/model/generator/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.generator;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
